+++ cypress-shared/integration/packing-optimization/2.1-rest-apis.spec.js
@@ -1,0 +1,14 @@
+import {
+  loginViaCookies,
+  preserveCookie,
+} from '../../support/common/support.js'
+import { pack } from '../../support/packing-optimization/api_testcase'
+import { packed } from '../../fixtures/pack.json'
+
+describe('Rest-api-testcases', () => {
+  loginViaCookies()
+
+  pack(packed)
+
+  preserveCookie()
+})